requirements.txt
Django==5.2.8
Pillow==12.0.0
django-crispy-forms==2.1
crispy-bootstrap4==1.0.2
gunicorn==23.0.0
python-decouple==3.8
